requirements.txt
pyvmomi
python-dotenv
tabulate